+++ nvimcom/src/apps/hover.c
@@ -12,7 +12,7 @@
 
 static int get_info(const char *s, char *p) {
     int i;
-    unsigned long nsz;
+    size_t nsz;
     const char *f[7];
     i = 0;
     while (i < 7) {
