+++ nvimcom/src/apps/utilities.h
@@ -1,6 +1,7 @@
 #ifndef UTILITIES_H
 #define UTILITIES_H
 
+#include <stddef.h>
 #ifdef WIN32
 #include <inttypes.h>
 #define bzero(b, len) (memset((b), '\0', (len)), (void)0)
@@ -13,7 +14,7 @@
 #define PRI_SIZET "zu"
 #endif
 
-char *grow_buffer(char **b, unsigned long *sz, unsigned long inc);
+char *grow_buffer(char **b, size_t *sz, size_t inc);
 void replace_char(char *s, char find, char replace);
 char *read_file(const char *fn, int verbose);
 char *esc_json(const char *input);
